feat(connectors): determine chain id by address if provider doesnt have a fetch chain id method
