+++ packages/client/src/connectors/ctrl.ts
@@ -5,12 +5,14 @@ import {
   type AddressType,
   BaseError,
   base64ToHex,
+  getAddressChainId,
   hexToBase64,
   MethodNotSupportedRpcError,
   ProviderNotFoundError,
   type SignPsbtParameters,
   UserRejectedRequestError,
 } from '@bigmi/core'
+import { ConnectorChainIdDetectionError } from '../errors/connectors.js'
 
 import { createConnector } from '../factories/createConnector.js'
 import type { UTXOConnectorParameters, UTXOWalletProvider } from './types.js'
@@ -146,7 +148,7 @@ export function ctrl(parameters: UTXOConnectorParameters = {}) {
       }
       try {
         const accounts = await this.getAccounts()
-        const chainId = await this.getChainId()
+        const chainId = getAddressChainId(accounts[0].address)
 
         if (!accountsChanged) {
           accountsChanged = this.onAccountsChanged.bind(this)
@@ -206,7 +208,15 @@ export function ctrl(parameters: UTXOConnectorParameters = {}) {
       throw new BaseError('Error getting accounts')
     },
     async getChainId() {
-      return chainId!
+      if (chainId) {
+        return chainId
+      }
+
+      const accounts = await this.getAccounts()
+      if (accounts.length === 0) {
+        throw new ConnectorChainIdDetectionError({ connector: this.name })
+      }
+      return getAddressChainId(accounts[0].address)
     },
     async isAuthorized() {
       try {
