--- packages/core/src/errors/config.ts
@@ -1,128 +1,0 @@
-import type { Address } from '../types/address.js'
-import type { ChainId, Network } from '../types/chain.js'
-
-import { BaseError } from './base.js'
-
-export type ChainNotConfiguredErrorType = ChainNotConfiguredError & {
-  name: 'ChainNotConfiguredError'
-}
-export class ChainNotConfiguredError extends BaseError {
-  override name = 'ChainNotConfiguredError'
-  constructor(chainId?: ChainId) {
-    super(`Chain ${chainId ? `: ${chainId}` : ''} not configured.`)
-  }
-}
-
-export type ConnectorAlreadyConnectedErrorType =
-  ConnectorAlreadyConnectedError & {
-    name: 'ConnectorAlreadyConnectedError'
-  }
-export class ConnectorAlreadyConnectedError extends BaseError {
-  override name = 'ConnectorAlreadyConnectedError'
-  constructor() {
-    super('Connector already connected.')
-  }
-}
-
-export type ConnectorNotConnectedErrorType = ConnectorNotConnectedError & {
-  name: 'ConnectorNotConnectedError'
-}
-export class ConnectorNotConnectedError extends BaseError {
-  override name = 'ConnectorNotConnectedError'
-  constructor() {
-    super('Connector not connected.')
-  }
-}
-
-export type ConnectorNotFoundErrorType = ConnectorNotFoundError & {
-  name: 'ConnectorNotFoundError'
-}
-export class ConnectorNotFoundError extends BaseError {
-  override name = 'ConnectorNotFoundError'
-  constructor() {
-    super('Connector not found.')
-  }
-}
-
-export type ConnectorAccountNotFoundErrorType =
-  ConnectorAccountNotFoundError & {
-    name: 'ConnectorAccountNotFoundError'
-  }
-export class ConnectorAccountNotFoundError extends BaseError {
-  override name = 'ConnectorAccountNotFoundError'
-  constructor({
-    address,
-    connector,
-  }: {
-    address: Address
-    connector: string
-  }) {
-    super(`Account "${address}" not found for connector "${connector}".`)
-  }
-}
-
-export type ConnectorChainMismatchErrorType = ConnectorAccountNotFoundError & {
-  name: 'ConnectorChainMismatchError'
-}
-export class ConnectorChainMismatchError extends BaseError {
-  override name = 'ConnectorChainMismatchError'
-  constructor({
-    connectionChainId,
-    connectorChainId,
-  }: {
-    connectionChainId: ChainId
-    connectorChainId: ChainId
-  }) {
-    super(
-      `The current chain of the connector (id: ${connectorChainId}) does not match the connection's chain (id: ${connectionChainId}).`,
-      {
-        metaMessages: [
-          `Current Chain ID:  ${connectorChainId}`,
-          `Expected Chain ID: ${connectionChainId}`,
-        ],
-      }
-    )
-  }
-}
-
-export type ConnectorNetworkMismatchErrorType =
-  ConnectorAccountNotFoundError & {
-    name: 'ConnectorNetworkMismatchError'
-  }
-export class ConnectorNetworkMismatchError extends BaseError {
-  override name = 'ConnectorNetworkMismatchError'
-  constructor({
-    connectionNetwork,
-    connectorNetwork,
-  }: {
-    connectionNetwork: Network
-    connectorNetwork: Network
-  }) {
-    super(
-      `The current network of the connector (${connectorNetwork}) does not match the connection's network (${connectionNetwork}).`,
-      {
-        metaMessages: [
-          `Current Network:  ${connectorNetwork}`,
-          `Expected Network: ${connectionNetwork}`,
-        ],
-      }
-    )
-  }
-}
-
-export type ConnectorUnavailableReconnectingErrorType =
-  ConnectorUnavailableReconnectingError & {
-    name: 'ConnectorUnavailableReconnectingError'
-  }
-export class ConnectorUnavailableReconnectingError extends BaseError {
-  override name = 'ConnectorUnavailableReconnectingError'
-  constructor({ connector }: { connector: { name: string } }) {
-    super(`Connector "${connector.name}" unavailable while reconnecting.`, {
-      details: [
-        'During the reconnection step, the only connector methods guaranteed to be available are: `id`, `name`, `type`, `uid`.',
-        'All other methods are not guaranteed to be available until reconnection completes and connectors are fully restored.',
-        'This error commonly occurs for connectors that asynchronously inject after reconnection has already started.',
-      ].join(' '),
-    })
-  }
-}
